// apps/web/src/generated/models/dto-user-access-token-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * commentPlayer API
 * ビデオ管理アプリケーションのREST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface DtoUserAccessTokenRequest {
    'password': string;
    'username': string;
}